\name{gidat}
\docType{data}
\alias{gidat}
\title{Data on NSAID and first GI bleed}
\description{
The data comprise ages in days at first gastro-intestinal (GI) bleed and treatments with non-steroidal anti-inflammatory drugs (NSAIDs). There are 838 simulated cases based on Tata et al (2005). 
}
\usage{gidat}
\format{A data frame containing 2920 rows and 6 columns. The column names are 'case' (individual identifier), 'sta' (age on the first day of the observation period), 'end' (age on last day of the observation period), 'bleed' (age at first GI bleed), 'ns' (age at first day of NSAID treatment episode), 'endns' (age at last day of NSAID treatment episode). The data are in format 'stack' with different exposure episodes on different rows.}
%\source{}
\references{Tata, L. J., Fortun, P. J., Hubbard, R. B., Smeeth, L., Hawkey,  C. J., Smith, P., Whitaker,  H. J., Farrington, C. P., Card, T. R., and West J. (2005). Does concurrent prescription of selective serotonin reuptake inhibitors and non-steroidal anti-inammatory drugs substantially increase the risk of upper gastrointestinal bleeding? Alimentary Pharmacology and Therapeutics 22, 175--181.
}
\keyword{datasets}
