\name{midat}
\docType{data}
\alias{midat}
\title{Data on respiratory tract infections and myocardial infarction}
\description{
These simulated data comprise ages in days at respiratory tract infection and first myocardial infarction in 1000 cases. 
}
\usage{midat}
\format{A data frame containing 2187 rows and 6 columns. The column names are 'case' (individual identifier), 'sta' (age on first day of the observation period), 'end' (age on last day of the observation period), 'mi' (age at first myocardial infarction), 'rti' (age at respiratory tract infection), 'cen' (the censoring indicator, equal to 1 if the observation period is censored, 0 if not). The data are in format "stack".}

%\source{}
\references{Smeeth L., Thomas, S. L.,  Hall, A. J., Hubbard, R., Farrington, P., Vallance, P. (2004). Risk of myocardial infarction and stroke after acute infection or vaccination. New England Journal of Medicine 351, 2611-2618.
}
\keyword{datasets}
