\name{sccs-data}
\Rdversion{1.1}
\alias{sccs-data}
\docType{data}
\title{
Standard Cross Cultural Sample dataset
}
\description{
}
\usage{data(sccs)}
\format{
}
\details{
}
%\source{
%%  ~~ reference to a publication or URL from which the data were obtained ~~
%}
\references{
}
\examples{
data(sccs)
}
\keyword{datasets}
