\name{siddat}
\docType{data}
\alias{siddat}
\title{Data on hexavalent vaccine and sudden infant death syndrome (SIDS)}
\description{
These simulated data comprise ages in days at hexavalent vaccination and SIDS in 300 cases. 
}
\usage{siddat}
\format{A data frame containing 300 rows and 5 columns. The column names are 'case' (individual identifier), 'sta' (age on first day of the observation period), 'end' (age on last day of the nominal observation period), 'sids' (age at SIDS), 'hex' (age at first hexavalent vaccination), 'hexd2' (age at second dose), 'hexd3' (age at third dose).}

%\source{}

\references{Kuhnert R., Hecker, H., Poethko-Muller, C., Schlaud, M., Vennemann, M., Whitaker, H.J., and Farrington C. P. (2011). A modified self-controlled case series method to examine association between multidose vaccinations and death. Statistics in Medicine 30, 666-677.
}
\keyword{datasets}
